@@ -18,16 +18,15 @@
     "@rails/ujs": "^7.0.4",
     "@types/babel__core": "7",
     "@types/node": "^18.15.11",
-    "@types/react": "^18.0.33",
-    "@types/react-dom": "^18.0.11",
-    "@types/webpack": "5",
+    "@types/react": "^19.1.9",
+    "@types/react-dom": "^19.1.7",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
+    "@vitejs/plugin-react": "^5.0.0",
     "alpine-turbo-drive-adapter": "^2.0.0",
     "alpinejs": "^2.2.5",
     "autoprefixer": "^10.4.20",
-    "axios": "^1.7.4",
-    "babel-loader": "8",
+    "axios": "^1.11.0",
     "babel-plugin-js-logger": "^1.0.17",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "classnames": "^2.3.1",
@@ -54,21 +53,21 @@
     "postcss-loader": "^7.2.4",
     "postcss-preset-env": "^8.0.1",
     "prop-types": "^15.8.1",
-    "puppeteer": "^18.0.0",
+    "puppeteer": "^24.16.1",
     "ramda": "^0.28.0",
     "rc-steps": "^6.0.0",
-    "react": "^18.2.0",
+    "react": "^18.3.1",
     "react-calendar": "2.18.1",
     "react-circular-progressbar": "^2.1.0",
     "react-datepicker": "^4.11.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-dropzone": "^14.2.3",
     "react-ga4": "^1.4.1",
     "react-infinite-scroll-component": "^6.1.0",
     "react-multi-select-component": "^4.3.4",
     "react-phone-number-input": "3.2.19",
     "react-responsive-carousel": "^3.2.23",
-    "react-router-dom": "^6.8.2",
+    "react-router-dom": "^7.8.0",
     "react-select": "^5.7.2",
     "react-table": "^7.8.0",
     "react-textarea-autosize": "^8.4.1",
@@ -85,9 +84,7 @@
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5",
     "uuid": "^9.0.0",
-    "webpack": "5",
     "webpack-assets-manifest": "5",
-    "webpack-cli": "4",
     "webpack-merge": "5",
     "webpack-sources": "^3.2.3",
     "world-countries": "^5.0.0",
@@ -97,7 +94,7 @@
     "@babel/eslint-parser": "^7.16.3",
     "@commitlint/cli": "^19.8.1",
     "@commitlint/config-conventional": "^19.8.1",
-    "@webpack-cli/serve": "^2.0.1",
+    "@vitejs/plugin-react-swc": "^4.0.0",
     "eslint": "^8.4.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.3",
@@ -110,9 +107,11 @@
     "eslint-plugin-unused-imports": "^2.0.0",
     "lefthook": "^1.12.2",
     "lint-staged": "^12.1.2",
+    "playwright": "^1.54.2",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
-    "webpack-dev-server": "4"
+    "vite": "^5.0.0",
+    "vite-plugin-ruby": "^5.1.0"
   },
   "lint-staged": {
     "app/**/*.{js,jsx,json,ts,tsx}": [
