@@ -6,15 +6,22 @@ import * as ReactRailsUJS from "react_ujs";
 
 import "./stylesheets/application.scss";
 
-require("alpinejs");
-require("jquery");
+import "alpinejs";
+import "jquery";
 
-require("settings");
+import "./settings";
 
 Rails.start();
 ActiveStorage.start();
 
 // Support component names relative to this directory:
-const componentRequireContext = require.context("src/components", true);
+// Note: require.context is webpack-specific and needs Vite alternative
+// Using dynamic imports for Vite compatibility
+const componentModules = import.meta.glob('./src/components/**/*.{js,jsx,ts,tsx}', { eager: true });
+const componentRequireContext = (path) => {
+  const fullPath = `./src/components${path.startsWith('./') ? path.slice(1) : '/' + path}`;
+  return componentModules[fullPath] || componentModules[fullPath + '.jsx'] || componentModules[fullPath + '.tsx'];
+};
+componentRequireContext.keys = () => Object.keys(componentModules).map(path => path.replace('./src/components', '.'));
 // eslint-disable-next-line react-hooks/rules-of-hooks
 ReactRailsUJS.useContext(componentRequireContext);
