@@ -3,16 +3,15 @@
 #
 # Table name: clients
 #
-#  id           :bigint           not null, primary key
-#  address      :string
-#  currency     :string           default("USD"), not null
-#  discarded_at :datetime
-#  email        :string
+#  id           :integer          not null, primary key
+#  company_id   :integer          not null
 #  name         :string           not null
+#  email        :string
 #  phone        :string
+#  address      :string
 #  created_at   :datetime         not null
 #  updated_at   :datetime         not null
-#  company_id   :bigint           not null
+#  discarded_at :datetime
 #  stripe_id    :string
 #
 # Indexes
@@ -20,13 +19,9 @@
 #  index_clients_on_company_id            (company_id)
 #  index_clients_on_discarded_at          (discarded_at)
 #  index_clients_on_email_and_company_id  (email,company_id) UNIQUE
-#  index_clients_on_email_trgm            (email) USING gin
+#  index_clients_on_email_trgm            (email)
 #  index_clients_on_name_and_company_id   (name,company_id) UNIQUE
-#  index_clients_on_name_trgm             (name) USING gin
-#
-# Foreign Keys
-#
-#  fk_rails_...  (company_id => companies.id)
+#  index_clients_on_name_trgm             (name)
 #
 
 # frozen_string_literal: true
