@@ -31,8 +31,7 @@ module MiruWeb
       g.test_framework :rspec, fixture: false
     end
 
-    # React configuration handled by Shakapacker
-    # config.react.server_renderer_extensions = ["jsx", "js", "tsx", "ts"]
+    # React configuration handled by Vite
 
     config.action_view.field_error_proc = Proc.new { |html_tag, instance|
       "#{html_tag}".html_safe
@@ -44,7 +43,7 @@ module MiruWeb
     config.action_mailer.postmark_settings = { api_token: ENV["POSTMARK_API_TOKEN"] }
 
     config.active_model.i18n_customize_full_message = true
-    # config.react.camelize_props = true # Handled by Shakapacker
+    # Props handling managed in React components
 
     # Use a real queuing backend for Active Job (and separate queues per environment).
     config.active_job.queue_adapter = :solid_queue
